[arcane,core] Déplace dans 'MeshMDVariableRef.h' les classes gérant les variables multi-dim du maillage.
