+++ arcane/src/arcane/utils/ArrayIndex.h
@@ -14,7 +14,7 @@
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-#include "arcane/utils/ArcaneGlobal.h"
+#include "arcane/utils/UtilsTypes.h"
 
 #include <array>
